% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoVVSURF.R
\name{CoVloo}
\alias{CoVloo}
\title{Construction of trees during leave-one-out procedure}
\usage{
CoVloo(X, prefix = NULL, mc.cores = 1, ...)
}
\arguments{
\item{X}{dataframe of input variables}

\item{prefix}{character string}

\item{mc.cores}{number of cores used for parallel computations}

\item{...}{passed to mclapply}
}
