% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoVVSURF.R
\name{covsurfloo}
\alias{covsurfloo}
\title{ClustOfVar combine with VSURF during leave-one-out procedure}
\usage{
covsurfloo(X, y, kval = 2:ncol(X), prefix = NULL, avail.tree = FALSE,
  covsurfs.object = NULL, path = NULL, nfor = 100, ...)
}
\arguments{
\item{X}{dataframe of input variables}

\item{y}{vector of responses}

\item{kval}{vector of number of classes to try}

\item{prefix}{character string}

\item{avail.tree}{logical value indicating if trees are already available}

\item{covsurfs.object}{character string}

\item{path}{path to covsurfs ojects}

\item{nfor}{number of forests to compute for averaged prediction}

\item{...}{passed to covsurf}
}

