% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoVVSURF.R
\name{plot.covsurf}
\alias{plot.covsurf}
\title{Plot the results from a covsurf object}
\usage{
\method{plot}{covsurf}(x, choice = "covsurf", ...)
}
\arguments{
\item{x}{an object of class "covsurf"}

\item{choice}{choice of the plot to perform: defaut is \code{"covsurf"}, other choices
are "vsurf", ...}

\item{...}{to be passed on plot function}
}
