% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoVVSURF.R
\name{predict.covsurf}
\alias{predict.covsurf}
\title{Make predictions for new data with a covsurf object}
\usage{
\method{predict}{covsurf}(object, newdata, select = TRUE, nfor = 1,
  mcores = 1)
}
\arguments{
\item{object}{an object of class "covsurf"}

\item{newdata}{data to predict}

\item{select}{if TRUE (default): predict only with selected variables,
else predict with all synthetic variables}

\item{nfor}{number of forests to build to obtain a collection of predictors}

\item{mcores}{number of cores used to parallel computations}
}
