% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoVVSURF.R
\name{repeated_simus}
\alias{repeated_simus}
\title{Helps to repeat experiments for simulated data}
\usage{
repeated_simus(n = 600, nrep = 50, nfor = 50, ...)
}
\arguments{
\item{n}{Number of observations in simulated data}

\item{nrep}{Number of repetitions of the simulations}

\item{nfor}{Number of forests prediction averaged}

\item{...}{Arguments passed to \code{\link{simus_covsurf}} function}
}
\description{
Helps to repeat experiments for simulated data
}
