% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{simu}
\alias{simu}
\title{Simulations of data}
\usage{
simu(n = 600, rho = 0.9, sigma = 1, name = "don", threshold = 0.5,
  seed = NULL)
}
\arguments{
\item{n}{number of observations}

\item{rho}{correlation between variables within a group}

\item{sigma}{standard deviation of the variables in the group of noise}

\item{name}{name of the data}

\item{threshold}{to binarize y}

\item{seed}{optional seed to pass to set.seed}
}
\value{
the matrix X of explonatory variables and y the binary variable to predict
}
