% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{simu_classif}
\alias{simu_classif}
\title{Simulations of data}
\usage{
simu_classif(n = 600, rho = 0.9, sigma = 1, threshold = 0.5,
  seed = NULL)
}
\arguments{
\item{n}{number of observations}

\item{rho}{correlation between variables within a group}

\item{sigma}{standard deviation of the variables in the group of noise}

\item{threshold}{to binarize y}

\item{seed}{optional seed to pass to set.seed}
}
\value{
the matrix X of explonatory variables and y the binary variable to predict
}
\references{
Combining clustering of variables and feature selection using random forests: the CoV/VSURF 
procedure, Marie Chavent, Robin Genuer, Jerome Saracco, hal-01345840
}
