% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{simu_reg}
\alias{simu_reg}
\title{Simulations of data}
\usage{
simu_reg(n = 600, rho = 0.9, sigma = 1, seed = NULL)
}
\arguments{
\item{n}{number of observations}

\item{rho}{correlation between variables within a group}

\item{sigma}{standard deviation of the variables in the group of noise}

\item{seed}{optional seed to pass to set.seed}
}
\value{
the matrix X of explonatory variables and y the numeric variable to predict
}
\references{
Combining clustering of variables and feature selection using random forests: the CoV/VSURF 
procedure, Marie Chavent, Robin Genuer, Jerome Saracco, hal-01345840
}
