% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoVVSURF.R
\name{simus_covsurf}
\alias{simus_covsurf}
\title{Perform covsurf on simulated datasets}
\usage{
simus_covsurf(data_name, data_test_name = "dontest", nfor = 100,
  Rdata.object = NULL, path = NULL, mcores = 1, ...)
}
\arguments{
\item{data_name}{name of the simulated dataset (the dataset will be loaded by 
\code{data(data_name)})}

\item{data_test_name}{name of the test dataset (the dataset will be loaded by 
\code{data(data_test_name)})}

\item{nfor}{number of forests for averaged predictions}

\item{Rdata.object}{optional Rdata name}

\item{path}{path to Rdata object}

\item{mcores}{number of cores used in parallel computations}

\item{...}{passed to VSURF}
}

