% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoVVSURF.R
\name{vsurfloo}
\alias{vsurfloo}
\title{VSURF during leave-one-out procedure}
\usage{
vsurfloo(X, y, vsurfs.object = NULL, path = NULL, nfor = 100, ...)
}
\arguments{
\item{X}{dataframe of input variables}

\item{y}{vector of responses}

\item{vsurfs.object}{optional VSURF objects}

\item{path}{path to VSURF objects}

\item{nfor}{number of forests for averaged predictions}

\item{...}{further parameters to be passed on VSURF function}
}

