% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoVVSURF.R
\name{covsurf}
\alias{covsurf}
\title{Combining ClustOfVar and VSURF}
\usage{
covsurf(X, y, kval = 2:ncol(X), tree = NULL, nse = 1, ...)
}
\arguments{
\item{X}{dataframe of input variables}

\item{y}{vector of responses}

\item{kval}{vector of number of classes to try}

\item{tree}{optional tree given by hclustvar}

\item{nse}{number of standard-deviation to add to select minimum of OOB rate}

\item{...}{passed to VSURF}
}
\examples{
data(don60)
kval <- c(2:15, seq(from = 20, to = ncol(X), by = 10))
don60covs <- covsurf(X, y, kval)
}

